Add files via upload
Write a program that takes two numbers n and k in the input and then takes n numbers from the input and stores them in an array. Your program should produce the sum of the numbers in each subsequence of this array. A number between 1 and n is read k times from the input as the end point of the sequence, and each time the sum of the numbers from the beginning of the array to that number is produced.